reorg code and bump Go to v1.22
