+++ internal/logger/init.go
@@ -15,19 +15,6 @@ type NewLoggerFunc func(w io.Writer, addSource bool, level slog.Leveler, color b
 // default to the console logger
 var CreateLogger NewLoggerFunc = NewConsole
 
-func SetLoggerFunc(name string) {
-	var loggers = map[string]NewLoggerFunc{
-		"console": NewConsole,
-		"json":    NewJSON,
-		"tint":    NewTinc,
-	}
-	var ok bool
-	CreateLogger, ok = loggers[name]
-	if !ok {
-		logger.Fatal("Invalid logger", "name", name)
-	}
-}
-
 // initialize the default logger to log to stderr and log at the warn level
 func init() {
 	w := os.Stderr
